+++ src/Compute/Compute.Test/ScenarioTests/VirtualMachineTests.cs
@@ -451,7 +451,7 @@ namespace Microsoft.Azure.Commands.Compute.Test.ScenarioTests
         {
             TestRunner.RunTestScript("Test-VMUserDataBase64Encoded");
         }
-
+        
         [Fact]
         [Trait(Category.AcceptanceType, Category.CheckIn)]
         public void TestVMNoPublicIPAddress()
@@ -459,5 +459,25 @@ namespace Microsoft.Azure.Commands.Compute.Test.ScenarioTests
             TestRunner.RunTestScript("Test-VMNoPublicIPAddress");
         }
 
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestVMForceDelete()
+        {
+            TestRunner.RunTestScript("Test-ForceDelete");
+        }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestVirtualMachineDiffDiskPlacement()
+        {
+            TestRunner.RunTestScript("Test-VirtualMachineDiffDiskPlacement");
+        }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestVirtualMachineHibernate()
+        {
+            TestRunner.RunTestScript("Test-VirtualMachineHibernate");
+        }
     }
 }