+++ src/KeyVault/KeyVault/Commands/Key/KeyRotationPolicy/InvokeAzKeyVaultKeyRotation.cs
@@ -9,7 +9,7 @@ namespace Microsoft.Azure.Commands.KeyVault.Commands.Key.KeyRotationPolicy
     /// <summary>
     /// Get the KeyRotationPolicy for the specified key in Key Vault.
     /// </summary>
-    [Cmdlet("Invoke", ResourceManager.Common.AzureRMConstants.AzurePrefix + "KeyVaultKeyRotation", SupportsShouldProcess = true, DefaultParameterSetName = ByVaultNameParameterSet)]
+    [Cmdlet(VerbsLifecycle.Invoke, ResourceManager.Common.AzureRMConstants.AzurePrefix + "KeyVaultKeyRotation", SupportsShouldProcess = true, DefaultParameterSetName = ByVaultNameParameterSet)]
     [OutputType(typeof(PSKeyVaultKey))]
     public class InvokeAzKeyVaultKeyRotation : KeyVaultOnlyKeyCmdletBase
     {
@@ -35,7 +35,10 @@ namespace Microsoft.Azure.Commands.KeyVault.Commands.Key.KeyRotationPolicy
         {
             NormalizeParameterSets();
 
-            WriteObject(this.Track2DataClient.RotateKey(VaultName, Name));
+            ConfirmAction(Name, Properties.Resources.RotateKey, () =>
+            {
+                WriteObject(this.Track2DataClient.RotateKey(VaultName, Name));
+            });
         }
     }
 }