+++ src/KeyVault/KeyVault/Models/KeyVaultDataServiceClient.cs
@@ -433,7 +433,7 @@ namespace Microsoft.Azure.Commands.KeyVault.Models
             throw new NotImplementedException();
         }
 
-        public PSKeyRotationPolicy UpdateKeyRotationPolicy(PSKeyRotationPolicy keyRotationPolicy)
+        public PSKeyRotationPolicy SetKeyRotationPolicy(PSKeyRotationPolicy keyRotationPolicy)
         {
             throw new NotImplementedException();
         }
@@ -2216,7 +2216,7 @@ namespace Microsoft.Azure.Commands.KeyVault.Models
             throw new NotImplementedException("Getting key rotation policies on managed HSM is only possible in track 2 SDK.");
         }
 
-        public PSKeyRotationPolicy UpdateManagedHsmKeyRotationPolicy(PSKeyRotationPolicy keyRotationPolicy)
+        public PSKeyRotationPolicy SetManagedHsmKeyRotationPolicy(PSKeyRotationPolicy keyRotationPolicy)
         {
             throw new NotImplementedException("Updating key rotation policies on managed HSM is only possible in track 2 SDK.");
         }