+++ src/Resources/ResourceManager/Implementation/Policy/SetAzurePolicyAssignment.cs
@@ -58,7 +58,7 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation
         /// Gets or sets the policy assignment not scopes parameter.
         /// </summary>
         [Parameter(Mandatory = false, ValueFromPipelineByPropertyName = true, HelpMessage = PolicyHelpStrings.SetPolicyAssignmentNotScopesHelp)]
-        [ValidateNotNullOrEmpty]
+        [ValidateNotNull]
         public string[] NotScope { get; set; }
 
         /// <summary>