+++ src/TrafficManager/TrafficManager/Models/TrafficManagerEndpoint.cs
@@ -49,6 +49,10 @@ namespace Microsoft.Azure.Commands.TrafficManager.Models
 
         public uint? MinChildEndpoints { get; set; }
 
+        public uint? MinChildEndpointsIPv4 { get; set; }
+
+        public uint? MinChildEndpointsIPv6 { get; set; }
+
         public List<string> GeoMapping { get; set; }
 
         public List<TrafficManagerIpAddressRange> SubnetMapping { get; set; }
@@ -62,6 +66,8 @@ namespace Microsoft.Azure.Commands.TrafficManager.Models
             endpoint.EndpointStatus = this.EndpointStatus;
             endpoint.GeoMapping = this.GeoMapping;
             endpoint.MinChildEndpoints = this.MinChildEndpoints;
+            endpoint.MinChildEndpointsIPv4 = this.MinChildEndpointsIPv4;
+            endpoint.MinChildEndpointsIPv6 = this.MinChildEndpointsIPv6;
             endpoint.Priority = this.Priority;
             endpoint.Target = this.Target;
             endpoint.TargetResourceId = this.TargetResourceId;