+++ src/Aks/Aks/ChangeLog.md
@@ -18,6 +18,10 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Fixed the typo in `New-AzAksCluster` [#16733]
+
+## Version 3.1.0
+* Added support of `load balancer` and `api server access` in `New-AzAksCluster` and `Set-AzAksCluster`. [#16575]
 
 ## Version 3.0.0
 * [Breaking Change] Updated parameter alias and output type of `Get-AzAksVersion`