+++ src/Storage/Storage.Management/ChangeLog.md
@@ -18,6 +18,13 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Fixed the issue that output number in console when update/copy blob sometimes [#16783]
+    -  `Set-AzStorageBlobContent` 
+    -  `Copy-AzStorageBlob` 
+* Updated help file, added more description for the asynchronous blob copy.
+    -  `Start-AzStorageBlobCopy`
+
+## Version 4.1.1
 * Fixed the failure of sync copy blob with long destination blob name [#16628]
     -  `Copy-AzStorageBlob` 
 
@@ -42,6 +49,24 @@
 * Supported create storage account with enable account level immutability with versioning, and create/update storage account with account level immutability policy.
     - `New-AzStorageAccount`
     - `Set-AzStorageAccount`
+* Supported AAD oauth storage context in storage table cmdlets.
+    - `Get-AzStorageCORSRule`
+    - `Get-AzStorageServiceLoggingProperty`
+    - `Get-AzStorageServiceMetricsProperty`
+    - `Get-AzStorageServiceProperty`
+    - `Get-AzStorageTable`
+    - `Get-AzStorageTableStoredAccessPolicy`
+    - `New-AzStorageTable`
+    - `New-AzStorageTableSASToken`
+    - `New-AzStorageTableStoredAccessPolicy`
+    - `Remove-AzStorageCORSRule`
+    - `Remove-AzStorageTableStoredAccessPolicy`
+    - `Set-AzStorageCORSRule`
+    - `Set-AzStorageServiceLoggingProperty`
+    - `Set-AzStorageServiceMetricsProperty`
+    - `Set-AzStorageServiceProperty`
+    - `Set-AzStorageTable`
+    - `Set-AzStorageTableStoredAccessPolicy`
 
 ## Version 3.12.0
 * Upgraded Azure.Storage.Blobs to 12.10.0